+++ CHANGELOG.md
@@ -14,12 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Upcoming
 ### Added
+- Add support for Hatch and uv-legacy style dependencies.
 - Add support for Python 3.13 -- 3.15
 - New status (!!) for source files that could not be parsed
 - Add CONTRIBUTING.md with contribution guidelines and coding standards
 - Fix README.md with correct usage of the CLI application
 ### Bugfix
-- Handle Python source files that declare a non-UTF8 encoding (PEP 263 encoding cookie or BOM).
+- Fix PEP621 parsing of development dependencies
+- Handle Python source files that declare a non-UTF8 encoding (PEP 263 encoding cookie or BOM)
 ## [0.13.0] - 2025-05-01 - Support for package - module mapping in pyproject.toml
 ### Added
 - Support for package - module mapping in pyproject.toml, to allow for different package and module names.